py filter imageFile null and close gate automatically

+++ camera_server.py
@@ -149,9 +149,6 @@ class MyHandler(http.server.BaseHTTPRequestHandler):
         cam = getIpCam(ip)
         if cam is None:
             return res
-        cam_in_out = "0"
-        if "in_out" in cam:
-            cam_in_out = cam["in_out"]
         if "open" in cam:
             open = cam["open"]
             #update cam with gate close
@@ -191,9 +188,6 @@ class MyHandler(http.server.BaseHTTPRequestHandler):
         cam = getIpCam(ip)
         if cam is None:
             return ""
-        cam_in_out = "0"
-        if "in_out" in cam:
-            cam_in_out = cam["in_out"]
         if serial_num == 0:
             cam_status[ip].carGio0 = serial_value
         else:
@@ -357,31 +351,34 @@ def updateCarInside(ip):
     reselt = data["AlarmInfoPlate"]["result"]["PlateResult"]
     size = "0"
     color = reselt["carColor"]
-    image_string = reselt["imageFile"]
     car_number = reselt["license"]
     file_name = "/storage/sdcard/Cars/" + car_number + ".png"
-    #add new car and open gate
-    if os.path.exists(file_name):
-        os.remove(file_name)
-    with open(file_name, "wb") as fh:
-        fh.write(base64.decodebytes(str.encode(image_string)))
-    addCarInside(car_number, "A", file_name, size, color)
+    
+    if "imageFile" in reselt:
+        image_string = reselt["imageFile"]
+        #add new car and open gate
+        if os.path.exists(file_name):
+            os.remove(file_name)
+        with open(file_name, "wb") as fh:
+            fh.write(base64.decodebytes(str.encode(image_string)))
+        addCarInside(car_number, "A", file_name, size, color)
 #update database after car go out
 def updateHistory(ip):
     data = cam_status[ip].carQueue.pop(0)
     reselt = data["AlarmInfoPlate"]["result"]["PlateResult"]
     car_number = reselt["license"]
-    image_string = reselt["imageFile"]
     file_name = "/storage/sdcard/Cars/" + car_number + ".png"
     backup_file_name = "/storage/sdcard/Cars_backup/" + nowTime.strftime("%Y_%m_%d_%H_%M_%S") + "_" + car_number + ".png"
     carInside = getCarInside(car_number)
     #close gate and show no available car slot
     addHistory(carInside, nowTime.strftime("%Y-%m-%d %H:%M:%S"), backup_file_name)
     deleteCarInside(car_number)
-    if os.path.exists(file_name):
-        os.remove(file_name)
-    with open(backup_file_name, "wb") as fh:
-        fh.write(base64.decodebytes(str.encode(image_string)))
+    if "imageFile" in reselt:
+        image_string = reselt["imageFile"]
+        if os.path.exists(file_name):
+            os.remove(file_name)
+        with open(backup_file_name, "wb") as fh:
+            fh.write(base64.decodebytes(str.encode(image_string)))
 #check the car can enter(check slots and cars_inside)
 def checkCanIn(data):
     ret = False
@@ -713,7 +710,7 @@ def run(server_class=http.server.HTTPServer, handler_class=MyHandler, port=8081)
     httpd.serve_forever()
     
 #initialize variables
-url = 'http://192.168.0.252:8080/function.php'
+url = 'http://localhost:8080/function.php'
 pic_dir = "/storage/sdcard/Cars"
 backup_dir = "/storage/sdcard/Cars_backup"
 os.makedirs(pic_dir, exist_ok=True)
